+++ src/services/App.php
@@ -47,6 +47,16 @@ class App extends Component
      */
     public $customers;
 
+    /**
+     * @var OrderStatuses
+     */
+    public $orderStatuses;
+
+    /**
+     * @var Emails
+     */
+    public $emails;
+
     public function init()
     {
         $this->settings = new Settings();
@@ -56,5 +66,7 @@ class App extends Component
         $this->subscriptions = new Subscriptions();
         $this->messages = new Messages();
         $this->customers = new Customers();
+        $this->orderStatuses = new OrderStatuses();
+        $this->emails = new Emails();
     }
 }